pyproject.toml
[tool.poetry]
name = "hydra-buddies"
version = "0.1.1"
description = "Un gestionnaire de configuration basé sur Hydra"
authors = ["MChrys <chrys.beltran@outlook.fr>"]
readme = "README.md"
packages = [{include = "hydra_buddies"}]
license = "MIT"
repository = "https://github.com/DemiurgeGalaxie/hydra-buddy"
keywords = ["hydra", "configuration", "omegaconf"]



[tool.poetry.dependencies]
python = "^3.12"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
click = "^8.1.7"
cookiecutter = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.poetry.scripts]
buddy = "hydra_buddies.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=hydra_buddies --cov-report=term-missing"


